requirements.txt
Flask
requests
beautifulsoup4
Pillow
gunicorn